requirements.txt
asgiref==3.3.1
beautifulsoup4==4.9.3
certifi==2020.12.5
chardet==4.0.0
colorama==0.4.4
configparser==5.0.2
coreapi==2.3.3
coreschema==0.0.4
crayons==0.4.0
Django==3.1.7
django-crontab==0.7.1
django-debug-toolbar==3.2.1
django-environ==0.4.5
django-filter==2.4.0
django-rest-swagger==2.2.0
django-simpleui==2021.4.3
djangorestframework==3.12.2
djangorestframework-camel-case==1.2.0
drf-yasg==1.20.0
environ==1.0
et-xmlfile==1.0.1
google-services-api==0.0.8
gunicorn==20.1.0
idna==2.10
ijson==3.1.4
inflection==0.5.1
itypes==1.2.0
jdcal==1.4.1
Jinja2==2.11.3
lxml==4.6.3
MarkupSafe==1.1.1
mysqlclient==2.0.3
numpy==1.19.5
openapi-codec==1.3.2
openpyxl==3.0.6
packaging==20.9
pandas==1.1.5
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
ruamel.yaml==0.17.4
ruamel.yaml.clib==0.2.2
selenium==3.141.0
simplejson==3.17.2
six==1.15.0
soupsieve==2.2.1
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.26.3
webdriver-manager==3.3.0
